+++ Libplanet/Store/StoreExtensions.cs
@@ -104,7 +104,19 @@ namespace Libplanet.Store
             }
 
             /// <inheritdoc/>
+            public void RenderError(IActionContext context, Exception exception)
+            {
+                // Does nothing.
+            }
+
+            /// <inheritdoc/>
             public void Unrender(IActionContext context, IAccountStateDelta nextStates)
+            {
+                // Does nothing.
+            }
+
+            /// <inheritdoc/>
+            public void UnrenderError(IActionContext context, Exception exception)
             {
                 // Does nothing.
             }